Rename tell_me_about_it to summarizer.
